CHORE : renamed auth to accounts in order to avoid naming conflicts with django.contrib.auth later
