+++ backend/accounts/views/register.py
@@ -5,7 +5,7 @@ from rest_framework import status
 from django.contrib.auth.tokens import default_token_generator
 from django.utils.http import urlsafe_base64_encode
 from django.utils.encoding import force_bytes
-from auth.serializers import RegisterSerializer
+from accounts.serializers import RegisterSerializer
 from django.core.mail import EmailMultiAlternatives
 from django.conf import settings
 from django.template.loader import render_to_string
