+++ backend/accounts/views/reset_password.py
@@ -4,7 +4,7 @@ from rest_framework import status
 from django.contrib.auth.tokens import default_token_generator
 from django.utils.http import urlsafe_base64_decode
 from django.contrib.auth.models import User
-from auth.serializers import ResetPasswordSerializer
+from accounts.serializers import ResetPasswordSerializer
 
 
 class ResetPasswordView(APIView):
