+++ backend/accounts/views/verify_email.py
