+++ backend/api/urls.py
@@ -7,6 +7,6 @@ router = routers.DefaultRouter()
 urlpatterns = [
     path("api/", include(router.urls)),
     path("api/admin", admin.site.urls),
-    path("api/auth/", include("auth.urls")),
+    path("api/auth/", include("accounts.urls")),
     path("api/<str:language_code>/", include("cards.urls")),
 ]
